+++ jest.config.js
@@ -1,5 +1,5 @@
 module.exports = {
   testEnvironment: 'node',
   collectCoverage: true,
-  collectCoverageFrom : ['./index.js']
+  collectCoverageFrom: ['./src'],
 };
